refactor(ui-service): integrate layer visualizer into UiConfigs editor chore(ui-service): improve analytics window help text
