--- Editor/UiLayerVisualizerWindow.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c0eaf32bf2ee84829b524c8318d1b5a6
